% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_IFC_features.R
\name{as_IFC_features}
\alias{as_IFC_features}
\title{Convert Features}
\usage{
as_IFC_features(x)
}
\arguments{
\item{x}{an object of class `IFCip_features`}
}
\value{
a list containing 2 members named:
-features, data.frame, object of class `IFC_features` 
-features_def, a list, object of class `IFC_features_def`
}
\description{
Function to convert an object of class `IFCip_features` as generated by ExtractFeatures() or ExtractBasic() 
to a list of `IFC_features` and `IFC_features_def`
}
