% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\title{Hu's Centroid}
\usage{
centroid(img)
}
\arguments{
\item{img}{an object of class `IFC_img` or `IFC_msk`.}
}
\value{
a named vector. where pix cx is x image's centroid and pix cy y image's centroid.
}
\description{
Hu's Centroid
}
\details{
See Visual pattern recognition by moment invariants authored
by Ming-Kuei Hu \doi{10.1109/TIT.1962.1057692}
}
