% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmoment.R
\name{cmoment}
\alias{cmoment}
\title{Hu's Central Moments}
\usage{
cmoment(img, centroid, p = 0, q = 0)
}
\arguments{
\item{img}{an object of class `IFC_img` or `IFC_msk`.}

\item{centroid}{a length 2 numeric vector of 'img' centroids.
When missing, the default, it will be computed from 'img'.}

\item{p}{an positive integer: p order. Default is 0.}

\item{q}{an positive integer: q order. Default is 0.}
}
\value{
a named vector, with atttributes:\cr
-order: p and q\cr
-matrix: the Hu's central moment matrix.
}
\description{
Hu's Central Moments
}
\details{
See Visual pattern recognition by moment invariants authored
by Ming-Kuei Hu \url{https://doi.org/10.1109/TIT.1962.1057692}
}
