% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_haralick.R
\name{compute_haralick}
\alias{compute_haralick}
\title{Haralick Texture}
\usage{
compute_haralick(img, msk, granularity = 3, bits = 4)
}
\arguments{
\item{img}{an object of class `IFC_img`, the image matrix.}

\item{msk}{an object of class `IFC_msk`, the mask matrix. If missing, the default mask information will be extracted from 'img'.
Otherwise, only elements superior to zero will be considered.}

\item{granularity}{a integer vector. Controls the grain of the texture. Default is 3. Allowed are [1-20].
For very fine textures, this value is small (1-3 pixels), while for very coarse textures, it is large (>10).}

\item{bits}{gray levels depth. Default is 4 (i.e. 2^4 = 16 gray levels). Allowed are [2-10].
Before applying compensation `IFC_img` have 2^12 = 4096 gray levels.}
}
\value{
array of Haralick's texture Features with dimensions corresponding to:\cr
-1st: the type of value returned,\cr
-2nd: the Haralick feature,\cr
-3rd: the granularity.
}
\description{
This function is designed to compute Haralick's Features.
}
\details{
See Haralick's original paper \url{https://haralick.org/journals/TexturalFeatures.pdf} and Löfstedt T. et al. url{https://doi.org/10.1371/journal.pone.0212110} used for the computation:
}
