% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_intensity.R
\name{compute_intensity}
\alias{compute_intensity}
\title{Intensity Computation}
\usage{
compute_intensity(img, msk)
}
\arguments{
\item{img}{an object of class `IFC_img`, the image matrix.}

\item{msk}{an object of class `IFC_msk`, the mask matrix. If missing, the default mask information will be extracted from 'img'.
Otherwise, only elements superior to zero will be considered.}
}
\value{
intensity wihtin 'msk'.
}
\description{
Computes intensity of image
}
