% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_perimeter.R
\name{compute_perimeter}
\alias{compute_perimeter}
\title{Perimeter Computation}
\usage{
compute_perimeter(msk, mag = 40, global = TRUE)
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{mag}{magnification. Default is 40. Allowed are 20, 40 and 60.}

\item{global}{whether to compute the perimeter globally or to evaluate the perimeter of each non 8-connected objects. Default is TRUE.}
}
\value{
a data.frame of 'msk' perimeter.
}
\description{
Computes perimeter of a mask
}
\details{
when global is TRUE the value returned is the total perimeter of all objects found in the mask.
If pixels from 2 different objects overlap, they are counted only once.
However, if FALSE, value returned is the number of perimeter of each single objects.
In case out border of 2 different objects overlap they are counted for each objects.
}
