% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shape.R
\name{compute_shape}
\alias{compute_shape}
\title{Shape Computation}
\usage{
compute_shape(msk, mag = 40)
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{mag}{magnification. Default is 40. Allowed are 20, 40 and 60.}
}
\value{
shape features from 'msk'.
}
\description{
Computes shape of a mask
}
