% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{compute_similarity}
\alias{compute_similarity}
\title{Images Similarity}
\usage{
compute_similarity(img1, img2, msk)
}
\arguments{
\item{img1}{a NumericMatrix, containing image values.}

\item{img2}{a NumericMatrix, containing image values.}

\item{msk}{a LogicalMatrix, containing mask. If missing, the default mask information will be extracted from 'img1'.
Otherwise, only elements superior to zero will be considered.}
}
\value{
similarity score within 'msk'.
}
\description{
This function is designed to score similarity between two images.
}
\details{
the similarity is the log transformed Pearson's Correlation Coefficient.
It is a measure of the degree to which two images are linearly correlated within a masked region.
}
