% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_background}
\alias{cpp_background}
\title{Image Background}
\arguments{
\item{img}{a NumericMatrix, containing image intensity values.}

\item{margin}{R_len_t number of rows margin used to compute background. Default is 4.}

\item{extra}{R_len_t number of extra columns used to compute background. Default is 0.}

\item{is_cif}{a bool whether 'ímg' originates from a cif_file or not. Default is false.}
}
\value{
a NumericVector of background mean and sd
}
\description{
This function is designed to compute image background on a "raw" image
}
\keyword{internal}
