% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_basic}
\alias{cpp_basic}
\title{Basic Features}
\arguments{
\item{img}{a NumericMatrix, containing image intensity values.}

\item{mag}{a double, magnification scale. Default is 1.0. Use:\cr
-1.0 for 20x\cr
-4.0 for 40x\cr
-9.0 for 60x.}
}
\value{
a NumericVector of basic features values.\cr
-Area: msk's area\cr
-circularity: msk's circularity\cr
-Minor Axis: msk's ellipsis minor axis\cr
-Major Axis: msk's ellipsis major axis\cr
-Aspect Ratio: msk's ratio of minor_axis over major_axis\cr
-Angle: msk's ellipsis angle with x axis (in radians)\cr
-theta: msk's ellipsis theta angle (in radians)\cr
-eccentricity: msk's ellipsis ecentricity\cr
-pix cx: msk's pixel x centroïd\cr
-pix cy: msk's pixel y centroïd\cr
-pix min_axis: msk's ellipsis minor axis in pixels\cr
-pix maj axis: msk's ellipsis major axis in pixels\cr
-pix count: msk's area in pixels\cr
-Raw Mean Pixel: img's mean pixel intensity\cr
-Raw Min Pixel: img's minimal pixel intensity\cr
-Raw Max Pixel: img's maximal pixel intensity.
}
\description{
This function is designed to compute very basic features based on Hu's moments + intensities.
}
\keyword{internal}
