% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_bbox}
\alias{cpp_bbox}
\title{Bounding Box of Convex Hull}
\arguments{
\item{pts}{a 2-column matrix defining the locations (x and y coordinates, respectively) of points.
It has to be an object of class `IFCip_convexhull`}

\item{scale}{a double used to scale the returned values.}
}
\value{
a NumericVector of features from convex hull.
}
\description{
Computes features from a Convex Hull
}
\keyword{internal}
