% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_bright_similarity}
\alias{cpp_bright_similarity}
\title{Images Bright Detail Similarity Measurement}
\arguments{
\item{img1}{a NumericMatrix, containing bright detail image values.}

\item{img2}{a NumericMatrix, containing bright detail image values.}

\item{msk}{a LogicalMatrix, containing mask.}
}
\value{
a double, the similarity.
}
\description{
This function is designed to score similarity between two bright detail images.
}
\details{
the bright detail similarity is the non-mean normalized version of the log transformed Pearson's Correlation Coefficient.
It is designed to compare the small bright image detail of two images within a masked region.\cr
See "Quantitative analysis of protein co-localization on B cells opsonized with rituximab and complement using the ImageStream multispectral imaging flow cytometer"
by P.V. Beum et al. Journal of Immunological Methods Volume 317, Issues 1–2, 20 December 2006, Pages 90-99 \doi{doi.org/10.1016/j.jim.2006.09.012}
}
\keyword{internal}
