% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_centroid}
\alias{cpp_centroid}
\title{Hu's Centroid}
\arguments{
\item{img}{a NumericMatrix, containing image intensity values.}
}
\description{
This function is designed to compute Hu's image centroids.
}
\keyword{internal}
