% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cmoment}
\alias{cpp_cmoment}
\title{Hu's Central Moment}
\arguments{
\item{img}{a NumericMatrix, containing image intensity values.}

\item{cx}{double, x centroid of the img\cr}

\item{cy}{double, y centroid of the img\cr}

\item{p}{uint8_t: p order. Default is 0.}

\item{q}{uint8_t: q order. Default is 0.}
}
\description{
This function is designed to compute Hu's image central moment.
}
\keyword{internal}
