% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cont}
\alias{cpp_cont}
\title{Image Contrast Enhancement}
\arguments{
\item{mat, }{a Matrix.}

\item{kernel, }{a Nullable Matrix.}
}
\value{
a Matrix of same type as 'mat'.
}
\description{
This function applies contrast enhancement on image.
}
\keyword{internal}
