% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_convexhull}
\alias{cpp_convexhull}
\title{Convex Hull}
\source{
Adaptation of Andrew's Monotone Chain Algorithm A. M. Andrew, 'Another Efficient Algorithm
for Convex Hulls in Two Dimension', Information Processing Letters, 9, 1979, pp216-219 reported
in M. A. Jayaram, Hasan Fleyeh, 'Convex Hulls in Image Processing: A Scoping Review', 
American Journal of Intelligent Systems, Vol. 6 No. 2, 2016, pp. 48-58.
}
\arguments{
\item{pts}{a 2-column matrix defining the locations (x and y coordinates, respectively) of points.
x has to be in column 0 and y in column 1 (C index, add 1 for R).}
}
\value{
a vector of row indices of 'pts' that constitutes convex hull vertices.
}
\description{
Computes 2D convex hull of a set of points.
}
\keyword{internal}
