% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_convolve2d}
\alias{cpp_convolve2d}
\title{Image Filtering by Convolution}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}
}
\value{
a NumericMatrix.
}
\description{
This function applies 2D convolution filtering on image.
}
\keyword{internal}
