% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cooc}
\alias{cpp_cooc}
\title{Haralick Co-occurrence Matrix}
\arguments{
\item{img}{a Rcpp::IntegerMatrix of class `IFCip_rescale`, containing image intensity values.}

\item{delta}{a Rcpp::IntegerVector of column and row shifts. If only one value is provided only row will be shifted.}
}
\value{
a Rcpp::IntegerMatrix Gray-Level Co-occurrence Matrices.
}
\description{
This function is designed to compute Haralick co-occurrence matrix
}
\details{
See 'Textural Features for Image Classification', Haralick et. al (1979),
available at: \url{https://haralick.org/journals/TexturalFeatures.pdf}
}
\keyword{internal}
