% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_ctl}
\alias{cpp_ctl}
\title{Contour Tracing Connected Component Labeling}
\arguments{
\item{mat}{a LogicalMatrix, containing mask.}

\item{global}{whether to compute the perimeter globally or to evaluate the perimeter of each non 8-connected objects. Default is false.
When true pixels of overlapping extra borders of objects are counted only once.}
}
\value{
a list whose members are:\cr
-matrix: an IntegerMatrix with connected component labels.\cr
-contours: an IntegerMatrix of identified contours, whose columns are x, y, label, direction and type.\cr
-nb_lab: the total number of components identified.
-perimeter: the number of pixels outside contours.
}
\description{
This function is designed to identify connected component.
}
\details{
adaptation of 'A linear-time component-labeling algorithm using contour tracing technique' from F. Chang, C.J. Chen and C.J. Lu.
Computer Vision and Image Understanding Volume 93, Issue 2, February 2004, Pages 206-220.\url{https://doi.org/10.1016/j.cviu.2003.09.002}
}
\keyword{internal}
