% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_dilate}
\alias{cpp_dilate}
\title{Image Dilatation}
\arguments{
\item{mat, }{a Matrix.}

\item{kernel, }{a Nullable Matrix.}
}
\value{
a Matrix of same type as 'mat'.
}
\description{
This function applies dilatation on image.
}
\details{
see 'Efficient 2-D grayscale morphological transformations with arbitrary flat structuring elements' from  E.R. Urbach, M.H.F. Wilkinson.
IEEE Transactions on Image Processing, 17(1):1-8, January 2008.\doi{10.1109/tip.2007.912582}
}
\keyword{internal}
