% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_dilate_old}
\alias{cpp_dilate_old}
\title{Brute Force Image Dilatation}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}

\item{iter, }{an uint8_t, number of time dilate should be iterated. Default is 0.}

\item{msk_, }{a NumericMatrix with finite values. Non-finite values will trigger an error. All non 0 values will be interpreted as true.
Default is R_NilValue, for using all 'mat' elements without masking anything.}
}
\value{
a NumericMatrix.
}
\description{
This function applies dilatation on image.
}
\details{
Brute force implementation now replaced by Urbach-Wilkinson algorithm.
}
\keyword{internal}
