% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_distance_eucl}
\alias{cpp_distance_eucl}
\title{Mask Euclidean Distance}
\arguments{
\item{msk}{an IntegerMatrix, containing connected components.}
}
\value{
an NumericMatrix.
}
\description{
This function is designed to compute Euclidean distance from background to centroïds' foreground
}
\keyword{internal}
