% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_distance_eucl_norm}
\alias{cpp_distance_eucl_norm}
\title{Mask Normalized Euclidean Distance}
\arguments{
\item{msk}{an IntegerMatrix, containing connected components.}
}
\value{
a NumericMatrix.
}
\description{
This function is designed to compute normalized Euclidean distance from background to centroïds' foreground
}
\keyword{internal}
