% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_distance_manh}
\alias{cpp_distance_manh}
\title{Mask Manhattan Distance}
\arguments{
\item{msk}{an IntegerMatrix, containing connected components.}
}
\value{
a NumericMatrix.
}
\description{
This function is designed to compute Manhattan distance from background to centroïds' foreground
}
\keyword{internal}
