% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_distance_manh_norm}
\alias{cpp_distance_manh_norm}
\title{Mask Normalized Manhattan Distance}
\arguments{
\item{msk}{an IntegerMatrix, containing connected components.}
}
\value{
an NumericMatrix.
}
\description{
This function is designed to compute normalized Manhattan distance from background to centroïds' foreground
}
\keyword{internal}
