% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_disttrans_eucl}
\alias{cpp_disttrans_eucl}
\title{Euclidean Distance Transform}
\arguments{
\item{img, }{a NumericMatrix.}
}
\value{
a NumericMatrix.
}
\description{
This function computes the Euclidean distance transform of an image by implementing A. Meijster algorithm.
}
\details{
adaptation of 'A General Algorithm For Computing Distance Transforms In Linear Time' from W.H. Hesselink, A. Meijster, J.B.T.M. Roerdink.
Mathematical Morphology and its Applications to Image and Signal Processing. February 2002, Pages 331-340.\doi{10.1007/0-306-47025-X_36}\cr
Values > 0 will be considered as foreground whereas all other values will be concidered as background (i.e. 0).
}
\keyword{internal}
