% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_erode}
\alias{cpp_erode}
\title{Image Erosion}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}

\item{iter, }{an uint8_t, number of time erode should be iterated. Default is 0.}
}
\value{
a NumericMatrix.
}
\description{
This function applies erosion on image.
}
\keyword{internal}
