% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_erode_ctl}
\alias{cpp_erode_ctl}
\title{Contours Erosion}
\arguments{
\item{ctl}{a List, containing contour tracing labeling, object of class `IFCip_ctl`}

\item{kernel, }{a NumericMatrix.}
}
\value{
a NumericMatrix.
}
\description{
This function applies contours erosion.
}
\keyword{internal}
