% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_features_hu2}
\alias{cpp_features_hu2}
\title{Hu's Full Features}
\arguments{
\item{img}{a NumericMatrix, containing image intensity values.}

\item{mag}{a double, magnification scale. Default is 1.0. Use:\cr
-1.0 for 20x\cr
-4.0 for 40x\cr
-9.0 for 60x.}
}
\value{
a NumericVector of Hu's moments values.\cr
-Area: img's area\cr
-circularity: img's circularity\cr
-Minor Axis: img's ellipsis minor axis\cr
-Major Axis: img's ellipsis major axis\cr
-Aspect Ratio: img's ratio of minor axis over major axis\cr
-Angle: img's ellipsis angle with x axis (in radians)\cr
-theta: img's ellipsis theta angle (in radians)\cr
-Eccentricity: img's ellipsis ecentricity\cr
-pix cx: img's pixel x centroïd\cr
-pix cy: img's pixel y centroïd\cr
-pix min axis: img's ellipsis minor axis in pixels\cr
-pix_maj_axis: img's ellipsis major axis in pixels\cr
-pix count: img's area in pixels\cr
-inv[1-7]: image invariant moments.
}
\description{
This function is designed to compute Hu's central moments + 7 invariant moments
}
\keyword{internal}
