% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_features_hu3}
\alias{cpp_features_hu3}
\title{Image Features Extraction}
\arguments{
\item{img}{a NumericMatrix, containing image intensity values.}

\item{msk}{an IntegerMatrix, containing msk components.}

\item{components}{an unsigned integer. Maximal component component number to retrieve features about.
Default is 0 to retrieve  features for all components.}

\item{mag}{a double, magnification scale. Default is 1.0. Use:\cr
-1.0 for 20x\cr
-4.0 for 40x\cr
-9.0 for 60x.}
}
\value{
a NumericMatrix whose rows are component numbers and columns are:\cr
-Area, area of the component\cr
-circularity, circularity of the component\cr
-Minor Axis, minor axis of the component\cr
-Major Axis, major axis of the component\cr
-Aspect Ratio, aspect ratio of the component\cr
-Angle, angle of the component\cr
-theta, theta of the component\cr
-eccentricity, eccentricity of the component\cr
-Minor Axis Intensity, intensity weighted minor axis of the component\cr
-Major Axis Intensity, intensity weighted major axis of the component\cr
-Aspect Ratio Intensity, intensity weighted aspect ratio of the component\cr
-Angle Intensity, intensity weighted angle of the component\cr
-theta intensity, intensity weighted theta of the component\cr
-eccentricity intensity, intensity weighted eccentricity of the component\cr
-pix cx, x centroid of the component in pixels\cr
-pix cy, y centroid of the component in pixels\cr
-pix min axis, minor axis of the component in pixels\cr
-pix maj axis, major axis of the component in pixels\cr
-pix count, number of pixels occupied by the component\cr
-inv[1-8], component Hu's invariant moments\cr
-Raw Mean Pixel, mean pixels intensity of the component\cr
-Raw Min Pixel, pixels intensity minimum of the component\cr
-Raw Max Pixel, pixels intensity maximum of the component\cr
-Std Dev, pixels intensity standard variation of the component\cr
-skewness, component's skewness\cr
-kurtosis, component's kurtosis.
}
\description{
This function is designed to compute image features.
}
\keyword{internal}
