% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_features_hu4}
\alias{cpp_features_hu4}
\title{Mask Features Extraction}
\arguments{
\item{msk}{an IntegerMatrix, containing msk components.}

\item{labels}{a Nullable IntegerVector corresponding to the desired label(s) to retrieve features about.
Default is \code{0} to retrieve features for all components.}

\item{mag}{a double, magnification scale. Default is \code{1.0}. Use:\cr
-1.0 for 20x\cr
-4.0 for 40x\cr
-9.0 for 60x.}
}
\value{
a NumericMatrix whose rows are component numbers and columns are:\cr
-Area, area of the component\cr
-circularity, circularity of the component\cr
-Minor Axis, minor axis of the component\cr
-Major Axis, major axis of the component\cr
-Aspect Ratio, aspect ratio of the component\cr
-Angle, angle of the component\cr
-theta, theta of the component\cr
-eccentricity, eccentricity of the component\cr
-pix cx, x centroid of the component in pixels\cr
-pix cy, y centroid of the component in pixels\cr
-pix min axis, minor axis of the component in pixels\cr
-pix maj axis, major axis of the component in pixels\cr
-pix count, number of pixels occupied by the component\cr
-Centroid Y, scaled Y centroid\cr
-Centroid X, scaled X centroid.
}
\description{
This function is designed to compute mask features.
}
\keyword{internal}
