% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fill}
\alias{cpp_fill}
\title{Contours Filling}
\arguments{
\item{ctl}{a List, containing contour tracing labeling, object of class `IFCip_ctl`}

\item{label}{an int corresponding to the label of desired set of contour to be filled.
Default is \code{0} to fill all sets of contours found.}

\item{i_border}{a bool, to whether or not draw inside contours if some were identified.}

\item{i_fill}{a bool, to whether or not fill inside contours if some were identified.}

\item{o_border}{a bool, to whether or draw external contours.}

\item{o_fill}{a bool, to whether or not fill external contours.}

\item{neg_border}{a bool, to whether or not border, if drawn, should be negated.}
}
\value{
an IntegerMatrix.
}
\description{
This function is designed to fill contours.
}
\keyword{internal}
