% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fill_default}
\alias{cpp_fill_default}
\title{Contours Default Filling}
\arguments{
\item{ctl}{a List, containing contour tracing labeling, object of class `IFCip_ctl`}

\item{labels}{a Nullable IntegerVector corresponding to the label(s) of desired set of contour to be filled.
Default is \code{0} to fill all sets of contours found.}

\item{i_neg_border}{a bool, to whether or not inside border, if drawn, should be negated. Default is \code{false}.}

\item{o_neg_border}{a bool, to whether or not external border, if drawn, should be negated. Default is \code{false}.}
}
\value{
an IntegerMatrix.
}
\description{
This function is designed to apply default contours filling.
}
\keyword{internal}
