% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fill_out}
\alias{cpp_fill_out}
\title{Contours Filling Outer Only}
\arguments{
\item{ctl}{a List, containing contour tracing labeling, object of class `IFCip_ctl`.}

\item{o_border}{a bool, to whether or draw external contours.}

\item{o_fill}{a bool, to whether or not fill external contours.}
}
\value{
an IntegerMatrix.
}
\description{
This function is designed to fill the most external contours.
}
\keyword{internal}
