% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_filter}
\alias{cpp_filter}
\title{Image Filtering}
\arguments{
\item{mat, }{a Matrix.}

\item{kernel, }{a Nullable Matrix.}

\item{method}{used for padding, a uint8_t. Default is \code{5}, allowed are [1-8].}

\item{k, }{constant used for padding, a double. Default is \code{NA_REAL}.}

\item{what, }{type of filtering, s std::string. Default is \code{""}.}
}
\value{
a NumericMatrix.
}
\description{
This function applies filtering on image.
}
\keyword{internal}
