% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_flip}
\alias{cpp_flip}
\title{Image Flip}
\arguments{
\item{mat}{NumericMatrix.}

\item{which}{bool. Default is true. Use true to flip horizontally and false vertically.}
}
\value{
a flipped matrix.
}
\description{
Function that flips mat
}
\keyword{internal}
