% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_floodfill}
\alias{cpp_floodfill}
\title{Connected Region Flood Filling}
\arguments{
\item{img, }{a NumericMatrix. The image to be modified.}

\item{markers, }{a NumericMatrix, It should be a matrix with at least 3 columns being "row", "col", and "value", respectively. It represents coordinates of the seeds to start filling 'img', with the new "value". Eventually, an additional column being "tolerance" can be provided.\cr
/!\ Note that "row" and "col" should be provided at C-level meaning 1st start at 0.}
}
\value{
a NumericMatrix, the modified image.
}
\description{
Flood fills image region.
}
\keyword{internal}
