% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_geo_tophat_black}
\alias{cpp_geo_tophat_black}
\title{Geodesic Black Top Hat}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}
}
\value{
a NumericMatrix.
}
\description{
This function applies geodesic black top hat on image.
}
\details{
see 'Morphological grayscale reconstruction in image analysis: applications and efficient algorithms' from  L. Vincent.
IEEE Transactions on Image Processing, 2(2):176-201, April 1993.\url{https://doi.org/10.1109/83.217222}\cr
Erosion in opening process is replaced by erosion reconstruction.
So, we have out = rec_erode(dilate(img)) - img.
}
\keyword{internal}
