% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_geo_tophat_white}
\alias{cpp_geo_tophat_white}
\title{Geodesic White Top Hat}
\arguments{
\item{img, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}
}
\value{
a NumericMatrix.
}
\description{
This function applies geodesic white top hat on image.
}
\details{
see 'Morphological grayscale reconstruction in image analysis: applications and efficient algorithms' from  L. Vincent.
IEEE Transactions on Image Processing, 2(2):176-201, April 1993.\doi{10.1109/83.217222}\cr
Dilation in closing process is replaced by dilation reconstruction.
So, we have out = img - rec_dilate(erode(img)).
}
\keyword{internal}
