% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_gradient}
\alias{cpp_gradient}
\title{Image Morphological Gradient}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}

\item{iter, }{an uint8_t, number of time dilate/erode should be iterated. Default is 0.}
}
\value{
a NumericMatrix.
}
\description{
This function applies morphological gradient on image.
}
\keyword{internal}
