% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_h_features}
\alias{cpp_h_features}
\title{Haralick Features}
\arguments{
\item{cooc}{a Rcpp::NumericMatrix of class `IFCip_cooc`, normalized co-occurrence matrix to compute Haralick's features from.}

\item{invariant}{a bool, whether to compute invariant Haralick's texture features. Default is false.
Not yet supported.}
}
\value{
a Rcpp::NumericVector of Haralick's texture features
}
\description{
This function is designed to compute Haralick's features
}
\details{
Haralick's invariant texture features are described in Löfstedt T, Brynolfsson P, Asklund T, Nyholm T, Garpebring A (2019) Gray-level invariant Haralick texture features.
PLoS ONE 14(2): e0212110. \url{https://doi.org/10.1371/journal.pone.0212110}
}
\keyword{internal}
