% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_k_sup_equal_M}
\alias{cpp_k_sup_equal_M}
\title{Matrix Logic Superior and Equal}
\arguments{
\item{mat}{a NumericMatrix}

\item{k}{constant to be checked. Default is 3.}
}
\value{
a logical matrix.
}
\description{
This function takes an NumericMatrix and checks for members superior or equal to k
}
\keyword{internal}
