% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_make_box}
\alias{cpp_make_box}
\title{Create a Box}
\arguments{
\item{size, }{a uint8_t of the desired kernel size.}
}
\value{
a LogicalMatrix.
}
\description{
This function is designed to create a box kernel.
}
\keyword{internal}
