% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_make_laplacian}
\alias{cpp_make_laplacian}
\title{Create Laplacian Kernel}
\arguments{
\item{size, }{a uint8_t of the desired kernel size.}

\item{sigma, }{a double, deviation of the filter used. Default is \code{-0.3}. If negative, \code{'sigma'} will be determined using \code{-1.0 * sigma * ((size - 1) * 0.5)}.}
}
\value{
a NumericMatrix.
}
\description{
This function is designed to create a laplacian kernel.
}
\keyword{internal}
