% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_median}
\alias{cpp_median}
\title{Image Median Filtering}
\arguments{
\item{mat, }{a Matrix.}

\item{kernel, }{a Nullable Matrix.}

\item{method}{used for padding, a uint8_t. Default is \code{5}, allowed are [1-8].}

\item{k, }{constant used for padding, a double. Default is \code{NA_REAL}.}
}
\value{
a NumericMatrix.
}
\description{
This function applies median filtering on image.
}
\keyword{internal}
