% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_padding}
\alias{cpp_padding}
\title{Image Padding}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}

\item{method, }{a uint8_t. Default is 1, allowed are [1-8].\cr
-1, extra cols / rows will be filled with 'k', returned 'out' will not be filled.\cr
-2, extra cols / rows will be filled with the closest col / row, returned 'out' will not be filled.\cr
-3, extra cols / rows will be filled mirroring neighbor cols / rows, returned 'out' will not be filled.\cr
-4, extra cols / rows will be filled repeating neighbor cols / rows, returned 'out' will not be filled.\cr
-5, extra cols / rows will be filled with 'k', returned 'out' will be filled with mat.\cr
-6, extra cols / rows will be filled with the closest col / row, returned 'out' will be filled with mat.\cr
-7, extra cols / rows will be filled mirroring neighbor cols / rows, returned 'out' will be filled with mat.\cr
-8, extra cols / rows will be filled repeating neighbor cols / rows, returned 'out' will be filled with mat.}

\item{k, }{a double, constant used when method is 1 or 4. Default is 0.0.}
}
\value{
a List whose elements are:\cr
-out, a NumericMatrix, with extra cols / rows\cr
-ori_c, a R_len_t with x coordinate of the 1st non extra element,\cr
-ori_r, a R_len_t with y coordinate of the 1st non extra element.
}
\description{
This function creates a new matrix with extra rows / cols according to input mat, kernel
}
\keyword{internal}
