% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_R_shift_M}
\alias{cpp_R_shift_M}
\title{Right Shift Matrix}
\arguments{
\item{mat}{a Rcpp::IntegerMatrix, containing image intensity values.}

\item{bits}{uint8_t number of bit to shift matrix values. Default is 4. Allowed are [2,10].
If shifted value does not respect [0, 2^bits - 1] an error is thrown.}
}
\value{
an Rcpp::IntegerMatrix.
}
\description{
This function is designed to right shift bits of a matrix to [0, 2^bits - 1]
}
\keyword{internal}
