% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_rescale_M}
\alias{cpp_rescale_M}
\title{hpp_rescale_M}
\arguments{
\item{img}{a Rcpp::IntegerMatrix, containing image intensity values.}

\item{msk_, }{a Rcpp::NumericMatrix with finite values. Non-finite values will trigger an error. All non 0 values will be interpreted as true.
Default is R_NilValue, for using all 'img' elements without masking anything.}

\item{bits}{uint8_t number of bit to shift matrix values. Default is 4. Allowed are [2,10].}
}
\value{
a Rcpp::IntegerMatrix.
}
\description{
This function is designed to rescale a matrix to [0, 2^bits - 1]
}
\keyword{internal}
