% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_rmoment}
\alias{cpp_rmoment}
\title{Hu's Raw Moment}
\arguments{
\item{img}{a NumericMatrix, containing image intensity values.}

\item{p}{uint8_t: p order. Default is 0.}

\item{q}{uint8_t: q order. Default is 0.}
}
\description{
This function is designed to compute Hu's image raw moment.
}
\keyword{internal}
