% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_scalerev}
\alias{cpp_scalerev}
\title{Image Reverse Scaling}
\arguments{
\item{img, }{a SEXP (logical, raw, integer or numeric) vector or matrix containing image intensity values.}

\item{sca_, }{a Rcpp::NumericVector of length 7 containing scaling information. Default is \code{R_NilValue} to use attr(img, "scale").}
}
\value{
a SEXP of same type as \code{'img'}
}
\description{
This function is designed to revert scaling of a SEXP
}
\keyword{internal}
