% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_sd}
\alias{cpp_sd}
\title{Image Standard Deviation Filtering}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}
}
\value{
a NumericMatrix.
}
\description{
This function applies standard deviation filtering on image.
}
\keyword{internal}
