% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_shift}
\alias{cpp_shift}
\title{Image Shift}
\arguments{
\item{mat}{a numeric matrix.}

\item{d_row}{an integer, giving row shift. Default is \code{0} for no change.}

\item{d_col}{an integer, giving col shift. Default is \code{0} for no change.}

\item{add_noise}{logical, if \code{true} adds normal noise when at least one new dimension is larger than original mat dimensions
Rcpp::rnorm() function is used. Default is \code{true}.}

\item{bg}{double, mean value of the background added if \code{'add_noise'} is \code{true}. Default is \code{0.0}.}

\item{sd}{double, standard deviation of the background added if \code{'add_noise'} is \code{true}. Default is \code{0.0}.}

\item{keep_size, }{a bool whether initial \code{'mat'} dimension should be kept. Default is \code{false}.}
}
\value{
a shifted matrix with additional rows/columns if \code{'d_row'} or \code{'d_col'} are different from \code{0}.
}
\description{
Function that shifts matrix rows and columns.
}
\keyword{internal}
