% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_similarity}
\alias{cpp_similarity}
\title{Images Similarity Measurement}
\arguments{
\item{img1}{a NumericMatrix, containing image values.}

\item{img2}{a NumericMatrix, containing image values.}

\item{msk}{a LogicalMatrix, containing mask.}
}
\value{
a double, the similarity.
}
\description{
This function is designed to score similarity between two images.
}
\details{
the similarity is the log transformed Pearson's Correlation Coefficient.
It is a measure of the degree to which two images are linearly correlated within a masked region.
}
\keyword{internal}
