% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_thinning_zs}
\alias{cpp_thinning_zs}
\title{Implementation of Zheng-Suen Thinning}
\arguments{
\item{mat}{a LogicalMatrix, containing mask.}
}
\value{
a LogicalMatrix with the mask thinned.
}
\description{
This function is designed to identify mask skeleton.
}
\details{
adaptation of 'A fast parallel algorithm for thinning digital patterns' from T. Y. Zhang, C. Y. Suen.
Communications of the ACM, March 1984 \url{https://doi.org/10.1145/357994.358023}.
}
\keyword{internal}
