% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_threshold}
\alias{cpp_threshold}
\title{Matrix Threshold}
\arguments{
\item{img}{a NumericMatrix}

\item{msk}{a NumericMatrix}

\item{k}{constant to be checked. Default is 0.}

\item{removal}{uint8_t, object removal method. Default is 0 for no removal. Otherwise, if\cr
-1, for clipped removal, keep non clipped foreground.\cr
-2, height clipped removal.\cr
-3, width clipped removal.\cr
-4, only keep background:.\cr
-5, only keep foreground.}
}
\value{
a logical matrix.
}
\description{
This function takes a image and checks for members superior or equal
to max(img) - k * diff(range(img)) / 100 within msk
}
\keyword{internal}
