% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_tophat_white}
\alias{cpp_tophat_white}
\title{Image White Top Hat}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}
}
\value{
a NumericMatrix.
}
\description{
This function applies white top hat on image.
}
\keyword{internal}
