% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_voronoi_manh}
\alias{cpp_voronoi_manh}
\title{Manhattan Vorono誰}
\arguments{
\item{img, }{a positive IntegerMatrix where values > 0 represent foreground seeds.}
}
\value{
an IntegerMatrix.
}
\description{
This function computes Vorono誰 diagram using Manhattan distance of a seed image.
}
\details{
img will be passed to connected component labelling and centro誰d of each identified components will be used to build Vorono誰 diagram.
}
\keyword{internal}
