% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_watershed}
\alias{cpp_watershed}
\title{Watershed Transformation}
\arguments{
\item{mat, }{a NumericMatrix; a distance transform matrix is expected.}

\item{connectivity, }{an unsigned short either 4 or 8 describing pixel neighborhood. Default is 8.}

\item{levels, }{an unsigned short determining the number elevation levels. Default is 256, should be at least 2.}
}
\value{
an IntegerMatrix.
}
\description{
This function computes the watershed transformation of an image.
}
\keyword{internal}
