% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_watershed_sv2}
\alias{cpp_watershed_sv2}
\title{Watershed Transformation SV2}
\source{
The algorithm is reviewed in 'The Watershed Transform: Definitions, Algorithms and Parallelization Strategies'
from Roerdink, J. B. T. M. and Meijster, A. (2000) in Fundamenta Informaticae, 41, 187-228 \url{https://doi.org/10.3233/FI-2000-411207}
}
\arguments{
\item{mat, }{a NumericMatrix; a distance transform matrix is expected.}

\item{connectivity, }{an uint8_t either 4 or 8 describing pixel neighborhood. Default is 8.}

\item{n_lev, }{an unsigned short determining the number of elevation levels. Default is 256, should be at least 2.}

\item{ws_draw, }{a bool; whether to draw watershed lines or not. Default is true.}

\item{ws_dilate}{, an uint8_t controlling watershed line expansion in pixel. Default is 0, for no expansion.
Then, increasing values will expand watershed lines by 2 pixels. This parameter only applies when 'ws_draw' is true.}
}
\value{
an IntegerMatrix.
}
\description{
This function computes the watershed transformation of an image.
}
\details{
adaptation of 'Watersheds in digital spaces: an efficient algorithm based on immersion simulations' from  L. Vincent and P. Soille.
In IEEE Transactions on Pattern Analysis and Machine Intelligence, 13(6):583-598, June 1991.\cr
}
\keyword{internal}
