% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_zernike1}
\alias{cpp_zernike1}
\title{Zernike's Features}
\source{
Adaptation from \url{https://github.com/aoles/EBImage} in v3.12.0, authored by Andrzej Oles, Gregoire Pau, Mike Smith, Oleg Sklyar, Wolfgang Huber, with contributions from Joseph Barry and Philip A. Marais \email{andrzej.oles@embl.de}.
}
\arguments{
\item{img}{a NumericMatrix, containing image intensity values.}

\item{cx}{a double. X centroid.}

\item{cy}{a double. Y centroid.}

\item{nmax}{a uint8_t, maximal order of Zernike polynomials to be computed. Default is 15. Values outside [0,99] will be clipped.
Be aware that computation of Zernike's moments can be quite long when 'nmax' is high.}

\item{radius}{a numeric, radius of the circle in pixels around object centers from which the features are calculated. Default is 15.}
}
\description{
This function is designed to compute Zernike's moments from image.
It will compute Zernike's moments but will not return image projection.
}
\keyword{internal}
