% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kernel.R
\name{make_kernel}
\alias{make_kernel}
\title{Create a Kernel}
\source{
'gaussian' is from EBImage::makeBrush R package and 'Laplacian' adapted from 'gaussian'
}
\usage{
make_kernel(size = 3, type = "box", sigma = 0.3)
}
\arguments{
\item{size}{the size of the desired kernel. Default is 3.\cr
It does not apply to 'type' "sobelx", "sobely", "scharr", "n4" nor "n8" which are of 'size' 3.}

\item{type}{a string indicating the type of the kernel Default is "box". Allowed are: 
"box", "cross", "plus", "disc", "diamond", "mean", "gaussian", "laplacian", "sobelx", "sobely", "scharr", "n4", "n8".}

\item{sigma}{standard deviation of the "gaussian" or "laplacian" 'type'. Default is 0.3.}
}
\value{
a kernel matrix.
}
\description{
Create a kernel of the desired 'type'.
}
\details{
"box", "cross", "plus", "disc", "diamond" will result in a logical matrix whereas the other will produce numeric a matrix.
}
