% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_clipped.R
\name{mask_clipped}
\alias{mask_clipped}
\title{Clipped Mask}
\usage{
mask_clipped(msk, borderLR = 1, borderTB = 6)
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{borderLR}{number of pixels considered as Left/Right edge.}

\item{borderTB}{number of pixels considered as Top/Bottom edge.}
}
\value{
a logical matrix of clipped components found.
}
\description{
This function is aimed to detect all components that are bound to edges.
Except if it is principal component
}
