% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_component.R
\name{mask_component}
\alias{mask_component}
\title{Component Mask}
\usage{
mask_component(msk, img, order = "Descending", feature = "Area")
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{img}{an object of class `IFC_img`, the image matrix.\cr
Can be missing when feature is "Area", "Aspect Ratio", "Circularity", "Thickness Max" or "Thickness Min".}

\item{order}{Either "Ascending" or "Descending". Default is "Descending".}

\item{feature}{Ranking feature. Default is "Area".\cr
Allowed are: "Area", "Aspect Ratio", "Bright Detail Intensity R3", "Bright Detail Intensity R7", "Bright Detail Similarity R3", "Circularity",
"Contrast", "Gradient RMS", "Intensity", "Max Pixel", "Mean Pixel", "Median Pixel", "Min Pixel", "Saturation Count", "Thickness Max", "Thickness Min".\cr
Only 'feature' = "Area" is possible for the moment}
}
\value{
list of mask(s) sorted according to 'order' and 'feature' selected.\cr
If no component is found (i.e. 'msk' is filled with 0), 'msk' is returned.
}
\description{
Apply component masking on an image.
}
