% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_dilate.R
\name{mask_dilate}
\alias{mask_dilate}
\title{Dilate Mask}
\usage{
mask_dilate(msk, n = 1)
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{n}{a positive integer corresponding to number of pixels to dilate. Default is 1.}
}
\value{
a logical matrix corresponding to the dilated area.
}
\description{
The dilate mask adds the selected number of pixels to all edges of the starting mask.
}
