% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_erode.R
\name{mask_erode}
\alias{mask_erode}
\title{Erode Mask}
\usage{
mask_erode(msk, n = 1)
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{n}{a positive integer corresponding to number of pixels to erode. Default is 1.}
}
\value{
a logical matrix corresponding to the eroded area.
}
\description{
The erode mask removes the selected number of pixels from all edges of the starting mask.
}
