% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_fill.R
\name{mask_fill}
\alias{mask_fill}
\title{Fill Mask}
\usage{
mask_fill(msk)
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}
}
\value{
a logical matrix corresponding to the filled area.
}
\description{
The fill mask fills in any holes in the starting mask.
}
