% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_identify.R
\name{mask_identify1}
\alias{mask_identify1}
\title{Identify Mask Version 1}
\usage{
mask_identify1(img, threshold = 0.95, size = 5)
}
\arguments{
\item{img}{the image matrix.}

\item{threshold}{quantile above which object(s) should be kept.}

\item{size}{the size of the kernel used to identify object(s). Default is 5. Should be higher than 3.}
}
\value{
an integer matrix of object(s) found.
}
\description{
The identify mask identifies objects within image using algorithm 1.
}
\keyword{internal}
