% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_identify.R
\name{mask_identify2}
\alias{mask_identify2}
\title{Identify Mask Version 2}
\usage{
mask_identify2(img, threshold = 0, size = 5)
}
\arguments{
\item{img}{the image matrix.}

\item{threshold}{standard deviation above which object(s) should be kept. Default is 0.}

\item{size}{the size of the kernel used to identify object(s). Default is 5. Should be higher than 3.}
}
\value{
an integer matrix of object(s) found.
}
\description{
The identify mask identifies objects within image using algorithm 2.
}
\keyword{internal}
