% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_range.R
\name{mask_range}
\alias{mask_range}
\title{Range Mask}
\usage{
mask_range(msk, area = c(0, 5000), aspect_ratio = c(0, 1))
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{area}{a length 2 numeric vector of area range defining the mask. Default is c(0,5000).}

\item{aspect_ratio}{a length 2 numeric vector of aspect ratio defining the mask. Default is c(0,1).}
}
\value{
a logical matrix of area(s) respecting range parameters found.
}
\description{
The range mask identifies region of desired area and aspect ration.
}
