% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_skeleton.R
\name{mask_skeleton}
\alias{mask_skeleton}
\title{Skeleton Mask}
\usage{
mask_skeleton(msk, option = "Thin")
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{option}{whether "Thin" or "Thick". Default is "Thin". Not yet implemented "Thick".}
}
\value{
a logical matrix corresponding to the skeleton area.
}
\description{
The skeleton mask provides the barebone structure of the object from the starting mask
}
