% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_threshold.R
\name{mask_threshold}
\alias{mask_threshold}
\title{Threshold Mask}
\usage{
mask_threshold(msk, img, threshold = 0, removal = 0)
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{img}{an object of class `IFC_img`, the image matrix.}

\item{threshold}{constant to be checked. Default is 0.}

\item{removal}{uint8_t, object removal method. Default is 0 for no removal. Otherwise, if\cr
-1, for clipped removal, keep non clipped foreground.\cr
-2, height clipped removal.\cr
-3, width clipped removal.\cr
-4, only keep background:.\cr
-5, only keep foreground.}
}
\value{
a logical matrix of area(s) above threshold found.
}
\description{
The threshold mask identifies intensity areas from an image that are superior or equal to max(img) - threshold * diff(range(img)) / 100 within msk.
}
