% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments_Hu.R
\name{moments_Hu}
\alias{moments_Hu}
\title{Hu's Moments}
\usage{
moments_Hu(img, mag = 40, full = TRUE)
}
\arguments{
\item{img}{an object of class `IFC_img` or `IFC_msk`.}

\item{mag}{magnification. Default is 40. Allowed are 20, 40 and 60.}

\item{full}{a logical. Controls the quantity of information that will be returned. Default is TRUE.
if full == TRUE, the eight first invariant moments will also be returned.}
}
\value{
a named vector of Hu's moments values.\cr
-Area: img's area\cr
-circularity: img's circularity\cr
-Minor Axis: img's ellipsis minor axis\cr
-Major Axis: img's ellipsis major axis\cr
-Aspect Ratio: img's ratio of minor_axis over major_axis\cr
-Angle: img's ellipsis angle with x axis (in radians)\cr
-theta: img's ellipsis theta angle (in radians)\cr
-eccentricity: img's ellipsis ecentricity\cr
-pix cx: img's pixel x centroïd\cr
-pix cy: img's pixel y centroïd\cr
-pix min axis: img's ellipsis minor axis in pixels\cr
-pix maj axis: img's ellipsis major axis in pixels\cr
-pix count: img's area in pixels\cr
-inv[1-7]: image invariant moments, depending on 'full' parameter.
}
\description{
This function is designed to compute Hu's central + invariant moments.
}
\details{
returned moments are based on Visual pattern recognition by moment invariants authored
by Ming-Kuei Hu \url{https://doi.org/10.1109/TIT.1962.1057692}
circularity is computed according to A Hu moment invariant as a shape circularity
measure authored by Joviša Žunić, Kaoru Hirota, Paul L. Rosin \url{https://doi.org/10.1016/j.patcog.2009.06.017}
}
