% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments_Zernike.R
\name{moments_Zernike}
\alias{moments_Zernike}
\title{Zernike's Moments}
\source{
adapted from File src/features_zernike.c in EBImage R package in its version 3.12.0 available at \url{https://www.bioconductor.org/packages/2.10/bioc/html/EBImage.html}
}
\usage{
moments_Zernike(img, centroid, nmax = 15, radius = 30, full = FALSE)
}
\arguments{
\item{img}{an object of class `IFC_img` or `IFC_msk`.}

\item{centroid}{a length 2 numeric vector of img centroids.
When missing, the default, it will be computed from img.}

\item{nmax}{an integer. Indicates the maximal order of Zernike polynomials
to be computed. Default value is 15. Values outside [0,99] will be clipped.
Be aware that computation of Zernike's moments can be quite long when 'nmax' is high.}

\item{radius}{a numeric. Defines the radius of the circle in pixels around
component center from which the features are calculated. Default is 30.}

\item{full}{a logical. Controls the quantity of information that will be returned. Default is FALSE.}
}
\value{
a named list zernike moments values.\cr
if full == TRUE, then Re and Im parts are also returned + zernike matrix projection of the image.
}
\description{
Computation of Zernike moment features from image.
}
\details{
Zernike features are computed by projecting image on Zernike complex polynomials.
}
