% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmoment.R
\name{rmoment}
\alias{rmoment}
\title{Hu's Raw Moments}
\usage{
rmoment(img, p = 0, q = 0)
}
\arguments{
\item{img}{an object of class `IFC_img` or `IFC_msk`.}

\item{p}{an integer: p order. Default is 0.}

\item{q}{an integer: q order. Default is 0.}
}
\value{
a named vector, wth atttributes:\cr
-order: p and q.\cr
-matrix: the Hu's raw moment matrix.
}
\description{
Hu's Raw Moments
}
\details{
See Visual pattern recognition by moment invariants authored
by Ming-Kuei Hu \url{https://doi.org/10.1109/TIT.1962.1057692}
}
