% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shape.R
\name{compute_shape}
\alias{compute_shape}
\title{Shape Computation}
\usage{
compute_shape(msk, mag = 40)
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{mag}{magnification. Default is 40. Allowed are 20, 40 and 60.}
}
\value{
shape features from 'msk'.
}
\description{
Computes shape of a mask
}
\details{
Based on contour tracing labeling by Chang F. et al. \url{https://doi.org/10.1016/j.cviu.2003.09.002},
convex hull computation from Andrew's Monotone Chain Algorithm \url{https://www.sciencedirect.com/science/article/abs/pii/0020019079900723}
and bounding box determination thanks to rotating calipers as described by Pirzadeh, Hormoz under supervision of Godfried T. Toussaint \url{https://escholarship.mcgill.ca/concern/theses/fx719p46g}
}
