% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cooc}
\alias{cpp_cooc}
\title{Haralick Co-Occurrence Matrix}
\arguments{
\item{img}{a Rcpp::IntegerMatrix of class `IFCip_rescale`, containing image intensity values.}

\item{msk}{a LogicalMatrix, containing mask.}

\item{delta}{uint8_t offset from which co-occurence has to be computed to. Default is 1.}
}
\value{
a list whose members are normalized Gray-Level Co-occurrence Matrices at angles 0, 45, 90 and 315.
}
\description{
This function is designed to compute Haralick co-occurrence matrix
}
\details{
See 'Textural Features for Image Classification', Haralick et. al (1979),
available at: \url{https://haralick.org/journals/TexturalFeatures.pdf}
}
\keyword{internal}
