% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_correlate2d}
\alias{cpp_correlate2d}
\title{Image Filtering by Correlation}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}

\item{method}{used for padding, a uint8_t. Default is 5 (with k = 0.0, see hpp_padding), allowed are [1-8].}
}
\value{
a NumericMatrix.
}
\description{
This function applies 2D correlation filtering on image.
}
\keyword{internal}
