% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_dilate}
\alias{cpp_dilate}
\title{Image Dilatation}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}

\item{iter, }{an uint8_t, number of time dilate should be iterated. Default is 0.}
}
\value{
a NumericMatrix.
}
\description{
This function applies dilatation on image.
}
\keyword{internal}
