% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_dilate_ctl}
\alias{cpp_dilate_ctl}
\title{Contours Dilatation}
\arguments{
\item{ctl}{a List, containing contour tracing labeling, object of class `IFCip_ctl`}

\item{kernel, }{a NumericMatrix.}

\item{iter, }{an uint8_t, number of time dilate should be iterated. Default is 0.}
}
\value{
a NumericMatrix.
}
\description{
This function applies contours dilatation.
}
\keyword{internal}
