% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_EMAX}
\alias{cpp_EMAX}
\title{Extended Maxima}
\arguments{
\item{img, }{a NumericMatrix.}

\item{img_min, }{a double. Minimal value of image range. Default is 0.0.
It is the value provided by the user, according to what the 'img' range should cover and not the actual minimal value of 'img'.}

\item{img_max, }{a double. Maximal value of image range. Default is 1.0.
It is the value provided by the user, according to what the 'img' range should cover and not the actual maximal value of 'img'.}

\item{kernel, }{a NumericMatrix; the structuring shape determining neighborhood. All non-zero elements will be considered as neighbors (except center).\cr
Default is R_NilValue, resulting in 8-connected pixels neighbors computation.}
}
\value{
a NumericMatrix of extended maxima of 'img'.
}
\description{
Mask the regional maxima of the corresponding h-maxima transformation.
}
\details{
see 'Morphological grayscale reconstruction in image analysis: applications and efficient algorithms' from  L. Vincent.
IEEE Transactions on Image Processing, 2(2):176-201, April 1993.\doi{10.1109/83.217222}\cr
EMAX is defined as RMAX(HMAX(img, h))
}
\keyword{internal}
