% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_erode_ctl}
\alias{cpp_erode_ctl}
\title{Contours Erosion}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}

\item{iter, }{an uint8_t, number of time erode should be iterated. Default is 0.}
}
\value{
a NumericMatrix.
}
\description{
This function applies contours erosion.
}
\keyword{internal}
