% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_gradient}
\alias{cpp_gradient}
\title{Image Morphological Gradient}
\arguments{
\item{mat, }{a Matrix.}

\item{kernel, }{a Nullable Matrix.}
}
\value{
a Matrix of same type as 'mat'.
}
\description{
This function applies morphological gradient on image.
}
\keyword{internal}
