% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_HMIN}
\alias{cpp_HMIN}
\title{H-Minima transformation}
\arguments{
\item{img, }{a NumericMatrix.}

\item{h, }{a double, specifying the minimal depth. Default is \code{NA_REAL}. When not \code{NA/NaN} it will be used instead of 'h_lev'}

\item{h_lev, }{an int, specifying the minimal depth normalized to n_lev (being h_lev out of n_lev). Default is \code{1}.}

\item{n_lev, }{an int determining the number levels used for 'img' rescaling. Default is 65536, should be at least 2.}

\item{kernel, }{a NumericMatrix; the structuring shape determining neighborhood. All non-zero elements will be considered as neighbors (except center).\cr
Default is R_NilValue, resulting in 8-connected pixels neighbors computation.}

\item{msk_, }{a Rcpp::NumericVector with finite values. Non-finite values will trigger an error. All non 0 values will be interpreted as true.
Default is R_NilValue, for using all 'img' elements without masking anything.}
}
\value{
a NumericMatrix of H-Minima transformation of 'img'.
}
\description{
Keep deepest valleys from image.
}
\details{
see 'Morphological grayscale reconstruction in image analysis: applications and efficient algorithms' from  L. Vincent.
IEEE Transactions on Image Processing, 2(2):176-201, April 1993.\doi{10.1109/83.217222}\cr
HMIN is the erosion reconstruction of (img + h) by kernel.
}
\keyword{internal}
