% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_laplacian}
\alias{cpp_laplacian}
\title{Image Laplacian}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}
}
\value{
a NumericMatrix.
}
\description{
This function applies Laplacian morphology on image.
}
\keyword{internal}
