% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_median}
\alias{cpp_median}
\title{Image Median Filtering}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}
}
\value{
a NumericMatrix.
}
\description{
This function applies median filtering on image.
}
\keyword{internal}
