% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mode}
\alias{cpp_mode}
\title{Image Mode Filtering}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}

\item{method}{used for padding, a uint8_t. Default is 5 (with k = NA_REAL, see hpp_padding), allowed are [1-8].}
}
\value{
a NumericMatrix.
}
\description{
This function applies mode filtering on image.
}
\keyword{internal}
