% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_padding}
\alias{cpp_padding}
\title{Image Padding}
\arguments{
\item{mat, }{a Matrix.}

\item{extra_rows, extra_cols}{number of extra rows and/or columns to add. Default is 0.}

\item{method, }{a uint8_t. Default is 1, allowed are [1-8].\cr
-1, extra cols / rows will be filled with 'k', returned 'out' will not be filled.\cr
-2, extra cols / rows will be filled with the closest col / row, returned 'out' will not be filled.\cr
-3, extra cols / rows will be filled mirroring neighbor cols / rows, returned 'out' will not be filled.\cr
-4, extra cols / rows will be filled repeating neighbor cols / rows, returned 'out' will not be filled.\cr
-5, extra cols / rows will be filled with 'k', returned 'out' will be filled with mat.\cr
-6, extra cols / rows will be filled with the closest col / row, returned 'out' will be filled with mat.\cr
-7, extra cols / rows will be filled mirroring neighbor cols / rows, returned 'out' will be filled with mat.\cr
-8, extra cols / rows will be filled repeating neighbor cols / rows, returned 'out' will be filled with mat.}

\item{k, }{a double, constant used when method is 1 or 5. Default is 0.0.}
}
\value{
a Matrix of same type as 'mat', with extra cols / rows
}
\description{
This function creates a new matrix with extra rows / cols according to input mat, kernel
}
\keyword{internal}
