% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_polydraw}
\alias{cpp_polydraw}
\title{Polygon Drawing}
\arguments{
\item{poly, }{a 2-column matrix defining the locations (x and y) of vertices of the polygon of interest.}

\item{border, }{an int used to trace polygon border.}

\item{fill, }{an int used to fill polygon.}

\item{mat_, }{an IntegerMatrix to be filled.}
}
\value{
copy of mat_ with poly
}
\description{
This function is designed to trace and fill polygon inside a matrix.
}
\keyword{internal}
