% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_rescale}
\alias{cpp_rescale}
\title{Image Scaling}
\arguments{
\item{img, }{a SEXP (logical, raw, integer or numeric) vector or matrix containing image intensity values.}

\item{msk_, }{a Rcpp::NumericVector with finite values. Non-finite values will trigger an error. All non 0 values will be interpreted as true.
Default is R_NilValue, for using all 'img' elements without masking anything.}

\item{value, }{a double; it is the replacement value that will be used when 'msk' element is false. Default is NA_REAL.}

\item{n_lev, }{an int determining the number of levels used for the computation. Default is 256.}

\item{invert, }{a bool determining whether 'img' should be scaled from min to max (when false, [min(img),max(img)] becoming [0,n_lev-1]) or inverted (when true, with [max(img),min(img)] rescaled to [0,n_lev-1]) values. Default is false.}

\item{bin, }{a bool determining whether 'img' should be binned or if scaling should be continuous. Default is true to return discrete values.}
}
\value{
a SEXP of same type as 'img' with class `IFCip_rescale`
}
\description{
This function is designed to scale a SEXP to [0, n_lev - 1]
}
\details{
when 'msk' is provided it has to be of the same dimensions as 'img', otherwise an error will be thrown.\cr
an error will be thrown also if 'msk' contains non-finite value.\cr
'img' range will be determined based on indices of non 0 'msk' values and only the values in 'img' at those indices will be scaled; the others will be filled with 'value'.
}
\keyword{internal}
