% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_rescale_M}
\alias{cpp_rescale_M}
\title{hpp_rescale_M}
\arguments{
\item{mat}{a Rcpp::NumericMatrix, containing image intensity values.}

\item{bits}{uint8_t number of bit to shift matrix values. Default is 4. Allowed are [2,10].
Rescaled values will normalized to [0, 2^bits - 1]}
}
\value{
an Rcpp::IntegerMatrix.
}
\description{
This function is designed to rescale a matrix to [0, 2^bits - 1]
}
\keyword{internal}
