% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_scalerev}
\alias{cpp_scalerev}
\title{Image Reverse Scaling}
\arguments{
\item{img, }{a SEXP (logical, raw, integer or numeric) vector or matrix containing image intensity values.}

\item{sca, }{a Rcpp::NumericVector of length 5 containing scaling information. Default is R_NilValue to use attr(img, "scale").}
}
\value{
a SEXP of same type as 'img'
}
\description{
This function is designed to revert scaling of a SEXP
}
\keyword{internal}
