% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_sd}
\alias{cpp_sd}
\title{Image Standard Deviation Filtering}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{kernel, }{a NumericMatrix.}

\item{method}{used for padding, a uint8_t. Default is \code{5}, allowed are [1-8].}

\item{k, }{constant used for padding, a double. Default is \code{NA_REAL}.}
}
\value{
a NumericMatrix.
}
\description{
This function applies standard deviation filtering on image.
}
\keyword{internal}
