% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_shift}
\alias{cpp_shift}
\title{Image Shift}
\arguments{
\item{mat}{a numeric matrix.}

\item{d_row}{an integer, giving row shift. Default is 0 for no change.}

\item{d_col}{an integer, giving col shift. Default is 0 for no change.}

\item{add_noise}{logical, if true adds normal noise when at least one new dimension is larger than original mat dimensions
Rcpp::rnorm() function is used. Default is true.}

\item{bg}{double, mean value of the background added if add_noise is true. Default is 0.}

\item{sd}{double, standard deviation of the background added if add_noise is true. Default is 0.}
}
\value{
a shifted matrix with additional rows/columns if d_row or d_col are different from 0.
}
\description{
Function that shifts mat according to d_row and d_col parameters
}
\keyword{internal}
