% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_tophat_self}
\alias{cpp_tophat_self}
\title{Image Self Complementary Top Hat}
\arguments{
\item{mat, }{a Matrix.}

\item{kernel, }{a Nullable Matrix.}
}
\value{
a Matrix of same type as 'mat'.
}
\description{
This function applies self complementary on image.
}
\keyword{internal}
