% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_watershed_sv1}
\alias{cpp_watershed_sv1}
\title{Watershed Transformation SV1}
\source{
MorphoLib plugin for ImageJ presents a Java implementation of the algorithm in  \url{https://github.com/ijpb/MorphoLibJ/blob/master/src/main/java/inra/ijpb/watershed/WatershedTransform2D.java} authored by Ignacio Arganda-Carreras
}
\arguments{
\item{mat, }{a NumericMatrix; a distance transform matrix is expected.}

\item{n_lev, }{an unsigned short determining the number of elevation levels. Default is 256, should be at least 2.}

\item{draw_lines, }{a bool; whether to draw watershed lines or not. Default is true.}

\item{invert, }{a bool; whether to fill from basins (lowest values) to peaks (highest values). Default is false.
When 'mat' is the result of the distance transformation of an image, peaks (highest values) represent largest distances from background.
Thus, they are the ones to be filled first; this is the default behavior with 'invert' set to false.}

\item{kernel, }{a NumericMatrix; the structuring shape determining neighborhood. All non-zero elements will be considered as neighbors (except center).\cr
Default is R_NilValue, resulting in 8-connected pixels neighbors computation.}

\item{msk_, }{a NumericMatrix with finite values. Non-finite values will trigger an error. All non 0 values will be interpreted as true.
Default is R_NilValue, for using all 'mat' elements without masking anything.}
}
\value{
an IntegerMatrix.
}
\description{
This function computes the watershed transformation of an image.
}
\details{
adaptation of 'Determining watersheds in digital pictures via flooding simulations' from P. Soille. and L. Vincent.
In Proc. SPIE 1360, Visual Communications and Image Processing '90: Fifth in a Series, (1 September 1990) \url{https://doi:10.1117/12.24211}.
}
\keyword{internal}
