% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{ExtractFeatures}
\alias{ExtractFeatures}
\title{Features Extraction}
\usage{
ExtractFeatures(
  ...,
  objects,
  offsets,
  removal = "masked",
  display_progress = TRUE,
  zmax = -1L,
  granularity = -1L,
  batch = 20,
  parallel = NULL
)
}
\arguments{
\item{...}{arguments to be passed to \code{\link{objectExtract}} with the exception of 'ifd' and 'bypass'(=TRUE).\cr
If 'param' is provided 'export'(="matrix"), 'mode'(="raw"), 'size'(="c(0,0)"), 'force_width'(="FALSE") and 'removal' will be overwritten.\cr
If 'offsets' are not provided extra arguments can also be passed with ... to \code{\link{getOffsets}}.\cr
/!\ If not any of 'fileName', 'info' and 'param' can be found in ... then attr(offsets, "fileName_image") will be used as 'fileName' input parameter to pass to \code{\link{objectParam}}.\cr
Remaining arguments with the exception of 'strategy' will be passed to \link[future]{plan}.}

\item{objects}{integers, indices of objects to use.
This argument is not mandatory, if missing, the default, all objects will be used.}

\item{offsets}{object of class `IFC_offset`. 
This argument is not mandatory but it may allow to save time for repeated image export on same file.}

\item{removal}{whether to compute features on "masked" object fo each individual channels or on the globally detected object "MC".
Allowed are "masked" or "MC". Default is "masked". Please note that it will overwrite 'param' value if provided.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{zmax}{maximal order of Zernike polynomials to be computed. Default is -1L for no computation.
Values outside [0,99] will be clipped. Be aware that computation of Zernike's Moments can be quite long when 'zmax' is high.}

\item{granularity}{an integer vector. Controls the grain of the Haralick texture.
Default is -1L for no computation. Allowed are [1-20].
For very fine textures, this value is small (1-3 pixels), while for very coarse textures, it is large (>10).}

\item{batch}{number of objects to process at the same time. Default is 20.}

\item{parallel}{whether to use parallelization. Default is NULL to use current \pkg{future}'s plan strategy.\cr
When FALSE, \link[future]{plan} will be called with sequential 'strategy'.
When TRUE, \link[future]{plan} will be called with either multisession 'strategy' on Windows or multicore otherwise..}
}
\value{
a 3D array of features values whose dimensions are [object, features, channel] of class `IFCip_features`.
}
\description{
Function to extract features from objects stored within rif and cif files.
}
\details{
arguments of objectExtract() from IFC package will be deduced from \code{\link{ExtractFeatures}} input arguments.
}
\examples{
if(!requireNamespace("IFCdata", quietly = TRUE)) {
  ## use a cif file
  file_cif <- system.file("extdata", package = "IFCdata", "example.cif")
  ## features extraction:
  ## the extraction is done for objects 1 to 50 only to allow example to run 
  ## in a reasonable amount of time and without parallelization to fulfill CRAN policies
  feat <- ExtractFeatures(fileName = file_cif, 
                          objects = 1:50,
                          display_progress = TRUE,
                          parallel = FALSE)
} else {
  message(sprintf('Please run `install.packages("IFCdata", repos = "\%s", type = "source")` \%s',
                  'https://gitdemont.github.io/IFCdata/',
                  'to install extra files required to run this example.'))
}
}
