% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_identify.R
\name{mask_identify}
\alias{mask_identify}
\title{Identify Mask}
\usage{
mask_identify(img, threshold = 0, size = 5, version = 2)
}
\arguments{
\item{img}{the image matrix.}

\item{threshold}{threshold above which object(s) should be kept. Default is 0.}

\item{size}{the size of the kernel used to identify object(s). Default is 5. Should be higher than 3.}

\item{version}{version to be used. Default is 2. Allowed are 1 and 2.}
}
\value{
an integer matrix of object(s) found.
}
\description{
The identify mask identifies objects within image.
}
\details{
With version 1 object identification is done by percentage thresholding on Laplacian of Gaussian after Gaussian denoising.\cr
Version 2 will perform thresholding on local standard deviation.
}
\keyword{internal}
