% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_skeleton.R
\name{mask_skeleton}
\alias{mask_skeleton}
\title{Skeleton Mask}
\usage{
mask_skeleton(msk, option = "Thin")
}
\arguments{
\item{msk}{an object of class `IFC_msk`, the mask matrix.}

\item{option}{whether "Thin" or "Thick". Default is "Thin". Not yet implemented "Thick".}
}
\value{
a logical matrix corresponding to the skeleton area.
}
\description{
The skeleton mask provides the barebone structure of the object from the starting mask
}
\details{
See 'A new thinning algorithm for binary images' from L. Ben Boudaoud, A. Sider, A. Tari.'
3rd international conference on control, engineering & information technology, May 2015. \url{https://doi.org/10.1109/CEIT.2015.7233099}.
}
